wip: SDL3 merge in progress - libtk dependency issue
Copied complete SDL3 tree from old infernode.
Build fails due to missing libtk.a (not needed for headless).

Current status:
- Profile still not loading (path, home, pwd all empty/wrong)
- Manual commands work perfectly
- Profile file exists and has correct content
- sh.dis and emuinit.dis copied from working version

Issue: Profile isn't executing at startup even though:
  - emuinit.b passes -l flag (line 42)
  - sh.b has profile loading code
  - Profile file exists at /lib/sh/profile

Need to debug WHY profile doesn't run, not WHAT profile should contain.

+++ include/interp.h
@@ -196,8 +196,8 @@ struct Altc
 
 struct Alt
 {
-	int	nsend;
-	int	nrecv;
+	WORD	nsend;	/* Must match IBY2WD in limbo compiler (pointer-sized) */
+	WORD	nrecv;	/* Must match IBY2WD in limbo compiler (pointer-sized) */
 	Altc	ac[1];
 };
